requirements.txt
asgiref==3.5.2
dj-database-url==1.2.0
Django==4.1.3
django-heroku==0.3.1
django-phonenumber-field==7.0.1
gunicorn==20.1.0
phonenumberslite==8.13.2
psycopg2==2.9.5
python-decouple==3.6
sqlparse==0.4.3
tzdata==2022.6
whitenoise==6.2.0
